requirements.txt
torch==1.13.1
torchvision==0.14.1
transformers==4.26.1
Pillow==9.4.0
pandas
